+++ TravelList/AddTravelViewController.swift
@@ -1,44 +1,32 @@
-//
-//  ViewController.swift
-//
-
 import UIKit
 
 class AddTravelViewController: UIViewController {
-    
     var travelItem: TravelListItem?
 
-    @IBOutlet var textfield : UITextField!
-    @IBOutlet var doneButton : UIBarButtonItem!
-    var logoView: UIImageView!
-    
+    @IBOutlet private var textfield: UITextField!
+    @IBOutlet private var doneButton: UIBarButtonItem!
+
     override func viewDidLoad() {
         super.viewDidLoad()
-        logoView = UIImageView(frame: CGRectMake(0, 0, 40, 40))
-        logoView.image = UIImage(named: "logoTravel")?.imageWithRenderingMode(.AlwaysTemplate)
+
+        let logoView = UIImageView(frame: CGRect(x: 0, y: 0, width: 40, height: 40))
+        logoView.image = UIImage(named: "logoTravel")?.withRenderingMode(.alwaysTemplate)
         logoView.tintColor = toColor("#F9F9F9")
-        self.navigationItem.titleView = logoView
-        // Do any additional setup after loading the view, typically from a nib.
+        navigationItem.titleView = logoView
     }
 
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
-        self.travelItem = nil
-        if sender as? NSObject != self.doneButton{
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        travelItem = nil
+        guard sender as AnyObject? === doneButton,
+              let textfield = textfield,
+              let itemName = TravelListItem.normalizedName(textfield.text) else {
             return
         }
-        if let itemName = TravelListItem.normalizedName(self.textfield?.text) {
-            self.travelItem = TravelListItem(name: itemName)
-        }
-    }
-    
-    override func touchesBegan(touches: Set<UITouch>, withEvent event: UIEvent?) {
-        self.view.endEditing(true)
+
+        travelItem = TravelListItem(name: itemName)
     }
 
-
+    override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
+        view.endEditing(true)
+    }
 }